refactor: split player events
