+++ src/player/events/rewind_event.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Event;
+
+#[derive(Event)]
+pub struct RewindEvent;
